[Add] - Cambio de roles y funcionalidad de org

+++ file_storage/convex/files.ts
@@ -19,15 +19,36 @@ export const generateUploadUrl = mutation(async (ctx) => {
 // This function checks if the user has access to the organization
 async function hasAccessToOrg(
   ctx: QueryCtx | MutationCtx,
-  tokenIdentifier: string,
   orgId: string
 ) {
-  const user = await getUser(ctx, tokenIdentifier);
+
+  const identity = await ctx.auth.getUserIdentity();
+
+    if (!identity) {
+      return null;
+    }
+
+  // Obtenemos el usuario actual
+  const user = await ctx.db
+    .query("users")
+    .withIndex("by_tokenIdentifier", (q) =>
+      q.eq("tokenIdentifier", identity.tokenIdentifier)
+    )
+    .first();
+
+    if (!user) {
+      return null;
+    }
+
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId);
 
-  return hasAccess;
+    if (!hasAccess) {
+      return null;
+    }
+
+  return {user };
 }
 
 // This is a mutation that creates a file in the databasem, is a table called "files"
@@ -43,16 +64,9 @@ export const createFile = mutation({
     error de prueba:
     throw new ConvexError("You must be logged in to upload a file"); */
 
-    // This is how you get the identity of the user that is logged con clerk, usa el archivo de auth.config.ts
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      throw new ConvexError("You must be logged in to upload a file");
-    }
 
     const hasAccess = await hasAccessToOrg(
       ctx,
-      identity.tokenIdentifier,
       args.orgId
     );
 
@@ -77,16 +91,9 @@ export const getFiles = query({
     favorites: v.optional(v.boolean()),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    // si no esta logeado, no puede ver los archivos
-    if (!identity) {
-      return [];
-    }
 
     const hasAccess = await hasAccessToOrg(
       ctx,
-      identity.tokenIdentifier,
       args.orgId
     );
 
@@ -107,22 +114,11 @@ export const getFiles = query({
     }
 
     if (args.favorites) {
-      // Obtenemos el usuario actual
-      const user = await ctx.db
-        .query("users")
-        .withIndex("by_tokenIdentifier", (q) =>
-          q.eq("tokenIdentifier", identity.tokenIdentifier)
-        )
-        .first();
-
-      if (!user) {
-        return files;
-      }
 
       const favorites = await ctx.db
         .query("favorites")
         .withIndex("by_userId_orgId_fileId", (q) =>
-          q.eq("userId", user._id).eq("orgId", args.orgId)
+          q.eq("userId", hasAccess.user._id).eq("orgId", args.orgId)
         )
         .collect();
 
@@ -156,6 +152,13 @@ export const deleteFile = mutation({
 
     if (!access) {
       throw new ConvexError("You do not have access to this file");
+    }
+
+    // Verificamos si el usuario es admin, aunque en frond no se vea por la proteccion de clerk esta es una forma de asegurarnos que no habran vulnerabilidades
+    const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("You do not have access to delete this file");
     }
 
     await ctx.db.delete(args.fileId);
@@ -197,17 +200,38 @@ export const toggleFavorite = mutation({
   },
 });
 
+// Obtiene todos los favoritos
+export const getAllFavorites = query({
+  args: {
+    orgId: v.string(),
+  },
+  async handler(ctx, args) {
+
+    const hasAccess = await hasAccessToOrg(ctx, args.orgId);
+
+    if (!hasAccess) {
+      return [];
+    }
+
+    // Buscamos si el archivo ya esta en favoritos
+    const favorites = await ctx.db
+      .query("favorites")
+      .withIndex("by_userId_orgId_fileId", (q) =>
+        q
+          .eq("userId", hasAccess.user._id)
+          .eq("orgId", args.orgId)
+      )
+      .collect();
+
+      return favorites;
+  },
+});
+
 // This function checks if the user has access to the file
 async function hasAccessToFile(
   ctx: QueryCtx | MutationCtx,
   fileId: Id<"files">
 ) {
-  const identity = await ctx.auth.getUserIdentity();
-
-  // si no esta logeado, no puede ver los archivos
-  if (!identity) {
-    return null;
-  }
 
   const file = await ctx.db.get(fileId);
 
@@ -217,7 +241,6 @@ async function hasAccessToFile(
 
   const hasAccess = await hasAccessToOrg(
     ctx,
-    identity.tokenIdentifier,
     file.orgId
   );
 
@@ -225,17 +248,5 @@ async function hasAccessToFile(
     return null;
   }
 
-  // Obtenemos el usuario actual
-  const user = await ctx.db
-    .query("users")
-    .withIndex("by_tokenIdentifier", (q) =>
-      q.eq("tokenIdentifier", identity.tokenIdentifier)
-    )
-    .first();
-
-  if (!user) {
-    return null;
-  }
-
-  return { file, user };
+  return { file, user: hasAccess.user };
 }
